refactor(todos): move the goal todo projection out of loopx/todos.py
loopx/todos.py sits exactly at its maintainability ceiling (2285 lines), so
threading one effective runtime root through the writer hooks cannot land
there without growing the module. Move the todo list projection block
(filtered summaries, the Markdown/event merge, and goal_todo_summaries) into
loopx/control_plane/todos/goal_todo_projection.py; list_goal_todos keeps its
behaviour and delegates to the new module.

Pure move: no call site outside loopx/todos.py used the moved helpers, and the
todo list, thin/explicit-limit, and agent-lane projection suites are unchanged.
